重新合并计算线程到一个文件
